Fix: Class "ContractLanguageEditor" has 27 methods, which is greater than 20 authorized. Split it into smaller classes.
