+++ app/Livewire/Contracts/ContractLanguageEditor.php
@@ -7,12 +7,20 @@ use App\Domains\Contract\Models\ContractClause;
 use App\Domains\Contract\Services\ContractService;
 use App\Domains\Contract\Services\ContractClauseService;
 use App\Domains\Core\Services\TemplateVariableMapper;
+use App\Livewire\Contracts\Traits\HandlesComments;
+use App\Livewire\Contracts\Traits\HandlesSearch;
+use App\Livewire\Contracts\Traits\HandlesUndoRedo;
+use App\Livewire\Contracts\Traits\ValidatesVariables;
 use Livewire\Component;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class ContractLanguageEditor extends Component
 {
     use AuthorizesRequests;
+    use HandlesComments;
+    use HandlesSearch;
+    use HandlesUndoRedo;
+    use ValidatesVariables;
 
     public Contract $contract;
     
@@ -252,7 +260,6 @@ class ContractLanguageEditor extends Component
             $metadata = $this->contract->metadata ?? [];
             $metadata['comments'] = $this->comments;
             
-            // Update contract with new content and variables
             $this->contract->update([
                 'content' => $this->editorMode === 'raw' ? $this->content : $this->previewContent,
                 'variables' => $this->variables,
@@ -274,88 +281,6 @@ class ContractLanguageEditor extends Component
         }
     }
     
-    // Search & Replace
-    public function search()
-    {
-        if (empty($this->searchQuery)) {
-            $this->searchResults = [];
-            return;
-        }
-        
-        $content = $this->editorMode === 'raw' ? $this->content : $this->previewContent;
-        $pattern = $this->useRegex 
-            ? '/' . $this->searchQuery . '/' . ($this->caseSensitive ? '' : 'i')
-            : '/' . preg_quote($this->searchQuery, '/') . '/' . ($this->caseSensitive ? '' : 'i');
-        
-        preg_match_all($pattern, $content, $matches, PREG_OFFSET_CAPTURE);
-        
-        $this->searchResults = $matches[0] ?? [];
-        $this->currentSearchIndex = 0;
-    }
-    
-    public function nextSearchResult()
-    {
-        if (!empty($this->searchResults)) {
-            $this->currentSearchIndex = ($this->currentSearchIndex + 1) % count($this->searchResults);
-        }
-    }
-    
-    public function previousSearchResult()
-    {
-        if (!empty($this->searchResults)) {
-            $this->currentSearchIndex = ($this->currentSearchIndex - 1 + count($this->searchResults)) % count($this->searchResults);
-        }
-    }
-    
-    public function replaceOne()
-    {
-        if (empty($this->searchResults) || !isset($this->searchResults[$this->currentSearchIndex])) {
-            return;
-        }
-        
-        $this->pushToUndoStack();
-        
-        $match = $this->searchResults[$this->currentSearchIndex];
-        $this->content = substr_replace($this->content, $this->replaceQuery, $match[1], strlen($match[0]));
-        
-        $this->hasChanges = true;
-        $this->search(); // Refresh search results
-        
-        $this->trackChange('replace_one', [
-            'from' => $match[0],
-            'to' => $this->replaceQuery,
-            'position' => $match[1],
-        ]);
-    }
-    
-    public function replaceAll()
-    {
-        if (empty($this->searchQuery)) {
-            return;
-        }
-        
-        $this->pushToUndoStack();
-        
-        $pattern = $this->useRegex 
-            ? '/' . $this->searchQuery . '/' . ($this->caseSensitive ? '' : 'i')
-            : '/' . preg_quote($this->searchQuery, '/') . '/' . ($this->caseSensitive ? '' : 'i');
-        
-        $count = 0;
-        $this->content = preg_replace($pattern, $this->replaceQuery, $this->content, -1, $count);
-        
-        $this->hasChanges = true;
-        $this->search(); // Refresh search results
-        
-        $this->trackChange('replace_all', [
-            'from' => $this->searchQuery,
-            'to' => $this->replaceQuery,
-            'count' => $count,
-        ]);
-        
-        session()->flash('success', "Replaced {$count} occurrence(s).");
-    }
-    
-    // Clause Reordering
     public function reorderClauses($newOrder)
     {
         if (!$this->canEdit) {
@@ -382,117 +307,6 @@ class ContractLanguageEditor extends Component
         $this->trackChange('reorder_clauses', ['new_order' => $newOrder]);
     }
     
-    // Undo/Redo
-    protected function pushToUndoStack()
-    {
-        $state = [
-            'content' => $this->content,
-            'variables' => $this->variables,
-            'includedClauses' => $this->includedClauses,
-            'timestamp' => now(),
-        ];
-        
-        array_push($this->undoStack, $state);
-        
-        // Keep stack size limited
-        if (count($this->undoStack) > $this->maxUndoSteps) {
-            array_shift($this->undoStack);
-        }
-        
-        // Clear redo stack when new action is performed
-        $this->redoStack = [];
-    }
-    
-    public function undo()
-    {
-        if (count($this->undoStack) <= 1) {
-            session()->flash('warning', 'Nothing to undo.');
-            return;
-        }
-        
-        // Move current state to redo stack
-        $currentState = array_pop($this->undoStack);
-        array_push($this->redoStack, $currentState);
-        
-        // Restore previous state
-        $previousState = end($this->undoStack);
-        $this->content = $previousState['content'];
-        $this->variables = $previousState['variables'];
-        $this->includedClauses = $previousState['includedClauses'];
-        
-        $this->hasChanges = true;
-        $this->generatePreview();
-        
-        session()->flash('success', 'Undone successfully.');
-    }
-    
-    public function redo()
-    {
-        if (empty($this->redoStack)) {
-            session()->flash('warning', 'Nothing to redo.');
-            return;
-        }
-        
-        $nextState = array_pop($this->redoStack);
-        array_push($this->undoStack, $nextState);
-        
-        $this->content = $nextState['content'];
-        $this->variables = $nextState['variables'];
-        $this->includedClauses = $nextState['includedClauses'];
-        
-        $this->hasChanges = true;
-        $this->generatePreview();
-        
-        session()->flash('success', 'Redone successfully.');
-    }
-    
-    // Comments
-    public function addComment($section = null)
-    {
-        if (empty($this->newComment)) {
-            return;
-        }
-        
-        $comment = [
-            'id' => uniqid(),
-            'section' => $section,
-            'content' => $this->newComment,
-            'author' => auth()->user()->name,
-            'created_at' => now()->toISOString(),
-            'resolved' => false,
-        ];
-        
-        array_push($this->comments, $comment);
-        $this->newComment = '';
-        $this->hasChanges = true;
-        
-        session()->flash('success', 'Comment added.');
-    }
-    
-    public function resolveComment($commentId)
-    {
-        foreach ($this->comments as &$comment) {
-            if ($comment['id'] === $commentId) {
-                $comment['resolved'] = true;
-                $comment['resolved_by'] = auth()->user()->name;
-                $comment['resolved_at'] = now()->toISOString();
-                $this->hasChanges = true;
-                break;
-            }
-        }
-        
-        session()->flash('success', 'Comment resolved.');
-    }
-    
-    public function deleteComment($commentId)
-    {
-        $this->comments = array_filter($this->comments, fn($c) => $c['id'] !== $commentId);
-        $this->hasChanges = true;
-        
-        session()->flash('success', 'Comment deleted.');
-    }
-    
-    // Change Tracking
     protected function trackChange($action, $details = [])
     {
         if (!$this->trackChanges) {
@@ -514,55 +328,6 @@ class ContractLanguageEditor extends Component
         }
     }
     
-    // Variable Validation
-    public function validateVariables()
-    {
-        $errors = [];
-        $warnings = [];
-        
-        // Check for missing required variables
-        foreach ($this->variables as $key => $value) {
-            if (empty($value) && $this->isVariableRequired($key)) {
-                $errors[] = "Variable '{$key}' is required but has no value.";
-            }
-        }
-        
-        // Check for unused variables
-        $content = $this->previewContent ?: $this->content;
-        foreach ($this->variables as $key => $value) {
-            if (strpos($content, '{{'.$key.'}}') === false) {
-                $warnings[] = "Variable '{$key}' is defined but not used in the contract.";
-            }
-        }
-        
-        // Check for undefined variables in content
-        preg_match_all('/\{\{([^}]+)\}\}/', $content, $matches);
-        foreach ($matches[1] as $match) {
-            $varName = trim($match);
-            // Skip conditional blocks and special syntax
-            if (strpos($varName, '#') === 0 || strpos($varName, '/') === 0 || strpos($varName, 'section') === 0) {
-                continue;
-            }
-            if (!isset($this->variables[$varName])) {
-                $errors[] = "Variable '{$varName}' is used but not defined.";
-            }
-        }
-        
-        return [
-            'valid' => empty($errors),
-            'errors' => $errors,
-            'warnings' => $warnings,
-        ];
-    }
-    
-    protected function isVariableRequired($key)
-    {
-        // Define required variables - this could be configured per template
-        $required = ['client_name', 'contract_date', 'service_provider_name'];
-        return in_array($key, $required);
-    }
-    
-    // Version Comparison
     public function compareWithVersion($versionId)
     {
         // This would load a previous version from database
